added RoomType table 	- defining room types (replacing the current hard coded ones) 	- room types are academic session independent added RoomTypeOption table 	- session dependent 	- relation (RoomType, Session) -> status (set of booleans), message 	- added status canScheduleEvents 		- use Room.getRoomType().getOption(Session).canScheduleEvents() to check whether it is possible to schedule events into a room of given type 	- these options (and messages) are editable via Edit Academic Session page

+++ JavaSource/org/unitime/timetable/solver/TimetableSolver.java
@@ -1155,7 +1155,7 @@ public abstract class TimetableSolver extends net.sf.cpsolver.coursett.Timetable
 		}
 	}
 	
-	public RoomReport getRoomReport(int startDay, int endDay, int nrWeeks, Integer roomType) {
+	public RoomReport getRoomReport(int startDay, int endDay, int nrWeeks, Long roomType) {
 		synchronized (currentSolution()) {
 			return new RoomReport((TimetableModel)currentSolution().getModel(), startDay, endDay, nrWeeks, roomType);
 		}
